Refactored end encounter conditions outside of Update
Refactored the code to remove many variables no longer needed in Encounter Control. Also created two methods for when the player wins and loses. Instead of checking every update to see if health has hit zero, the two methods are attached to the event system PlayerDeath. This is invoked when the health hits zero, which is only checked whenever health is altered.