Deleted non-mouse mode and added E pressed delegate
Removed all code that used arrows to traverse the player hand. This will not be used and instead we will stick to mouse controls. Added an E press delegate that includes all the code for updating deck visuals. The sound manager also adds a method that plays the reload method to the delegate.