Refactored to remove references to MusicManager and WeaponMono
Extracted code in EncounterControl that called MusicManager methods and instead put those methods in the MusicManager class. These methods are aded to the delegates "start" and "end" and are invoked at the correct time by Encounter Control

Similarly extracted code to be put in WeaponMono. Also moved all time slot spawning and despawning to WeaponMono where it adds appropiate methods to start and end delegates.