Main menu finished
Altered Menu Manager to not be time based. Instead, used animation events. The animations that the buttons played call the Menu Manager's method when a specific frame is reached. With this, the timings of sounds and method calls can be specific. The main menu buttons now work and call the correct methods.