Bug fixes
Resized two cards for consistentcy.
Also, realized that the player bullet count was not resetting after combat. Added a method to do that and Bullet Manager adds that method to the start delegate.
Also, sometimes hitboxes wouldn't be disabled. This was fixed by disabling the defend hitboxes when combat starts. The method disableHitbox is added to start delegate by Defense Manager.